Add :S to save current selection into Clipboard
Atm it works well *if* vim is running. If it is not running then
attempting to paste the content of the + register into any X app will
result in nothing and `xclip -sel clip -o` returns `target STRING is not
available`.

To circumvent this you can use the :S command to save to current content
of + into the clipboard X register, before leaving vim or pausing it (^C
or ^Z).