Refactor Find command, add FindAll [vim]
`:Find` now uses `git ls-files` and filters its output according to a
regex passed to the local search engine (either `rg` or `grep`).
`:FindAll` is the old version of this function which works the same way
but uses `find` instead. Every path listed in the wildignore is ignored
by `:FindAll`.